backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
aiohttp==3.9.3
beautifulsoup4==4.12.3
openai==1.12.0
pandas==2.2.1
numpy==1.26.4
python-jose==3.3.0
pytest==8.0.2
pytest-asyncio==0.23.5 